% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapapps.R
\docType{package}
\name{snapapps}
\alias{snapapps}
\alias{snapapps-package}
\title{snapapps: Collection of SNAP Shiny apps.}
\description{
The \code{snapapps} package contains a collection of open source Shiny apps by Matthew Leonawicz at the
Scenarios Network for Alaska and Arctic Planning, University of Alaska Fairbanks.
It is a member package in the apps and docs sector of the SNAPverse collection of R packages.
Many of the apps in this package originate from \url{https://github.com/ua-snap/shiny-apps}.
Additional apps can be found in other SNAPverse apps and docs sector packages.
Related sector packages include \code{snapdash}, \code{snapflex} and \code{snapdocs}.
}
