% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapapps.R
\name{snapp}
\alias{snapp}
\title{Run SNAP Shiny app examples}
\usage{
snapp(id, source = "local", local_mode = "normal")
}
\arguments{
\item{id}{character, the name of the application. See \link{snapps} for available IDs.}

\item{source}{character, source of app. See details.}

\item{local_mode}{character, display mode.}
}
\description{
Launch a Shiny app in your browser.
}
\details{
The \code{source} defaults to \code{"local"} for package apps. If set to \code{"remote"},
\code{snapp} will launch the app in your browser using the canonical url rather than the package app.
Options for \code{local_mode} are \code{"normal"} (default) or \code{"showcase"}
for tandem reactive code highlighting, and metadata if applicable. \code{local_mode} applies to \code{source = "local"} package apps.
}
\examples{
\dontrun{snapp("rv4")}
}
