% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapapps.R
\name{snapp_resources}
\alias{snapp_resources}
\title{Get a snapapps resource path}
\usage{
snapp_resources(type = "images")
}
\arguments{
\item{type}{character, resource type. \code{"images"} is the only currently available resource type.}
}
\description{
Get a resource path in an app to local package resources.
}
\details{
This function is typically called in an app that has been imported into the \code{snapapps} package.
While external apps may not share all common resources, once they are moved into the package it can be more efficient to have
only a single copy of a resource such as an image file and alter the code of each relevant app to point to that package file.

This function is only used during development to allow included apps to share common resources.
For applicable apps, you will find a call to this function near the top of \code{ui.R}, which would not be found in a
version of the app outside of this package.
}
\examples{
\dontrun{snapp_resources()}
}
