% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapapps.R
\name{snapps}
\alias{snapps}
\title{Basic metadata for all apps in snapapps}
\usage{
snapps()
}
\value{
a data frame.
}
\description{
This function returns a data frame with basic meta data for all Shiny apps in \code{snapapps}.
This includes app ID, name, short description, canonical remote url, redirect url and an indication of whether the app
is complete or partial with respect to the canonical app.
For example, did the local app require available data sets to be reduced or removed for inclusion in \code{snapapps}?
}
\examples{
snapps()
}
