% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapapps.R
\name{snapp}
\alias{snapp}
\title{Run SNAP Shiny app examples}
\usage{
snapp(id)
}
\arguments{
\item{id}{character, the name of the application. See details.

There are X appsavailable maps:
\describe{
\item{\code{rv1}}{Distributions of random variables v1}
\item{\code{rv2}}{Distributions of random variables v2}
\item{\code{rv3}}{Distributions of random variables v3}
\item{\code{rv4}}{Distributions of random variables v4}
}}
}
\description{
Launch a Shiny app in your browser.
}
\details{
Available apps include:
}
\examples{
\dontrun{snapp("rv4")}
}
